Refactor testsuites to match convention
Move IO tests under Test.IO so it's obvious where to find stuff.
